Move processing to garden (WIP)
