+++ etl/steps/data/garden/iea/2024-07-04/critical_minerals.py
@@ -415,8 +415,140 @@ def run(dest_dir: str) -> None:
     for table in [tb_demand, tb_supply]:
         table["process"] = table["process"].replace({"Mining": "Mine", "Refining": "Refinery"})
 
+    # Create a table for global demand.
+    # For each case-scenario-mineral-year, we need the global demand of all technologies.
+    # NOTE: Global mineral demand, including uses outside clean tech, is only given for "Base case" for a few minerals (for which "Other uses" is explicitly given).
+    minerals_with_total_demand = tb_demand[tb_demand["technology"] == "Other uses"]["mineral"].unique().tolist()
+    tb_demand_global = (
+        tb_demand[(tb_demand["case"] == "Base case") & (tb_demand["mineral"].isin(minerals_with_total_demand))]
+        .groupby(["case", "scenario", "mineral", "process", "year"], observed=True, as_index=False)
+        .agg({"demand": "sum"})
+        .rename(columns={"demand": "global_demand"})
+    )
+
+    # Create a table for global supply.
+    tb_supply_global = (
+        tb_supply[tb_supply["country"] == "World"]
+        .drop(columns=["country"], errors="raise")
+        .rename(columns={"supply": "global_supply"})
+    )
+
+    # Prepare supply by country table.
+    def create_supply_by_country_table(tb_supply: Table, tb_demand_global: Table, tb_supply_global: Table) -> Table:
+        tb_supply = tb_supply.copy()
+
+        # Add global supply.
+        tb_supply = tb_supply.merge(tb_supply_global, on=["case", "year", "mineral", "process"], how="left")
+        assert set(tb_supply[tb_supply["supply"] == tb_supply["global_supply"]]["country"]) == set(["World"])
+
+        # We assume that all demand is of refined minerals.
+        tb_supply = tb_supply.merge(tb_demand_global, on=["case", "year", "mineral", "process"], how="left")
+
+        # Add "share" columns.
+        tb_supply["supply_as_a_share_of_global_demand"] = 100 * tb_supply["supply"] / tb_supply["global_demand"]
+        tb_supply["supply_as_a_share_of_global_supply"] = 100 * tb_supply["supply"] / tb_supply["global_supply"]
+
+        # Drop unnecessary columns.
+        tb_supply = tb_supply.drop(columns=["global_demand", "global_supply"], errors="raise")
+
+        # After merging with global demand, "scenario" is added to the supply table.
+        # Supply does not change depending on the scenario, but demand does, and therefore supply as a share of global demand changes.
+        # Therefore, fill all cases where "scenario" is now nan (for "Mine" process) with some reasonable label.
+        # TODO: Later, when flattening the table, assert that all scenarios for total supply yield the same values, and assign the same label to scenario.
+        ALL_SCENARIOS_LABEL = "All scenarios"
+        tb_supply["scenario"] = tb_supply["scenario"].astype("string").fillna(ALL_SCENARIOS_LABEL)
+
+        return tb_supply
+
+    def create_demand_by_technology_table(tb_demand: Table, tb_demand_global: Table, tb_supply_global: Table) -> Table:
+        tb_demand = tb_demand.copy()
+
+        # From the original supply table, we need the global supply of each mineral-year (given only for the "Base case").
+        # TODO: For consistency with the current minerals explorer, we should rename them "Mine" and "Refinery".
+
+        # Add total demand to the demand table.
+        tb_demand = tb_demand.merge(tb_demand_global, on=["case", "scenario", "mineral", "process", "year"], how="left")
+        # TODO: For "Graphite (all grades: natural and synthetic)", "Other uses" has all the demand. There may be a bug.
+        #  t[t["demand"]==t["global_demand"]]
+        # TODO: After fixing that, consider renaming to "Graphite" and add a subtitle or footnote.
+
+        # Add total supply to the demand table.
+        tb_demand = tb_demand.merge(tb_supply_global, on=["case", "year", "mineral", "process"], how="left")
+
+        # Add "share" columns.
+        tb_demand["demand_as_a_share_of_global_demand"] = 100 * tb_demand["demand"] / tb_demand["global_demand"]
+        tb_demand["demand_as_a_share_of_global_supply"] = 100 * tb_demand["demand"] / tb_demand["global_supply"]
+
+        return tb_demand
+
+    tb_supply = create_supply_by_country_table(
+        tb_supply=tb_supply, tb_supply_global=tb_supply_global, tb_demand_global=tb_demand_global
+    )
+    tb_demand = create_demand_by_technology_table(
+        tb_demand=tb_demand, tb_supply_global=tb_supply_global, tb_demand_global=tb_demand_global
+    )
+
+    def create_demand_by_technology_flat(tb_demand_by_technology: Table) -> Table:
+        # Create a wide-format table.
+        tb_demand_by_technology_flat = tb_demand_by_technology.pivot(
+            index=["technology", "year"],
+            columns=["mineral", "process", "case", "scenario"],
+            values=["demand", "demand_as_a_share_of_global_demand", "demand_as_a_share_of_global_supply"],
+            join_column_levels_with="|",
+        )
+        # Adapt table to grapher.
+        tb_demand_by_technology_flat = tb_demand_by_technology_flat.rename(columns={"technology": "country"})
+
+        # Remove empty columns.
+        tb_demand_by_technology_flat = tb_demand_by_technology_flat.dropna(axis=1, how="all")
+
+        return tb_demand_by_technology_flat
+
+    # Create a wide-format table of demand by technology.
+    tb_demand_by_technology_flat = create_demand_by_technology_flat(tb_demand_by_technology=tb_demand)
+
+    def create_supply_by_country_flat(tb_supply_by_country: Table) -> Table:
+        # Create a wide-format table.
+        tb_supply_by_country_flat = tb_supply_by_country.pivot(
+            index=["country", "year"],
+            columns=["mineral", "process", "case", "scenario"],
+            values=["supply", "supply_as_a_share_of_global_demand", "supply_as_a_share_of_global_supply"],
+            join_column_levels_with="|",
+        )
+
+        # Remove empty columns.
+        tb_supply_by_country_flat = tb_supply_by_country_flat.dropna(axis=1, how="all")
+
+        return tb_supply_by_country_flat
+
+    # Create a wide-format table of supply by country.
+    tb_supply_by_country_flat = create_supply_by_country_flat(tb_supply_by_country=tb_supply)
+
+    # Remove "World" from supply data (since all countries should always add up to World).
+    # TODO: Add sanity check for this.
+    tb_supply_by_country_flat = tb_supply_by_country_flat[tb_supply_by_country_flat["country"] != "World"].reset_index(
+        drop=True
+    )
+
+    # Improve metadata.
+    # TODO: Do this properly.
+    for table in [tb_demand_by_technology_flat, tb_supply_by_country_flat]:
+        for column in table.columns:
+            table[column].m.title = column
+            if "_share_of_" in column:
+                table[column].m.unit = "%"
+                table[column].m.short_unit = "%"
+            else:
+                table[column].m.unit = "tonnes"
+                table[column].m.short_unit = "t"
+
+    # Improve its format.
+    # NOTE: This should be done after improving metadata, otherwise titles will get snake-cased.
+    tb_demand_by_technology_flat = tb_demand_by_technology_flat.format(short_name="demand_by_technology")
+    tb_supply_by_country_flat = tb_supply_by_country_flat.format(short_name="supply_by_country")
+
     # Format output tables conveniently.
-    tb_supply = tb_supply.format(["case", "country", "year", "mineral", "process"])
+    tb_supply = tb_supply.format(["case", "country", "year", "mineral", "process", "scenario"])
     tb_demand = tb_demand.format(["case", "year", "mineral", "process", "scenario", "technology"])
 
     #
@@ -424,7 +556,10 @@ def run(dest_dir: str) -> None:
     #
     # Create a new garden dataset with the same metadata as the meadow dataset.
     ds_garden = create_dataset(
-        dest_dir, tables=[tb_demand, tb_supply], check_variables_metadata=True, default_metadata=ds_meadow.metadata
+        dest_dir,
+        tables=[tb_demand, tb_supply, tb_demand_by_technology_flat, tb_supply_by_country_flat],
+        check_variables_metadata=True,
+        default_metadata=ds_meadow.metadata,
     )
 
     # Save changes in the new garden dataset.
